Move AM_INIT_AUTOMAKE macro and require automake 1.11 or later.
My understanding is that AM_INIT_AUTOMAKE should be called before
calling other AM_* macros, though what we have appears to work I've
moved the AM_INIT_AUTOMAKE macro earlier in the file as I believe this
is correct.

I've also make automake 1.11 or later a requirement, this is still
pretty old, but is needed to support AM_SILENT_RULES which I'm going
to add in a later patch.

ChangeLog:

	* configure.ac: Move AM_INIT_AUTOMAKE earlier in the file, require
	automake 1.11 or later.
	* configure: Regenerate.