Add support for Wall and Werror, turn these on by default.
This patch adds the Wall and Werror flags into the CFLAGS used to
build the tests.  These flags can be turned off at configure time
using the flags '--disable-werror' and '--disable-build-warnings'.
Currently, almost every test fails to build due to errors and/or
warnings, so these flags will be needed in order to get BEEBS to
build.

ChangeLog:

	* configure.ac: Add support for Wall and Werror.
	* src/common.mk.am: Add the warning and error flags into the
	AM_CFLAGS.
	* Makefile.in: Regenerate.
	* configure: Regenerate.
	* doc/Makefile.in: Regenerate.
	* src/*/Makefile.in: Regenerate.