Fix compiler warnings in crc32.
Missing include and unused variables.  The unused variable
check_output has been commented out rather than removed as the TODO
comment suggests that there is some other issue to investigate in this
test.
When we revisit the TODO comment later the commented out code will
either be fixed and re-enabled, or removed.

ChangeLog:

	* src/crc32/crc_32.c: Add include for stdlib.h.
	(main): Comment out check_output, cast defined but unused
	variable 'output' to void.