The default --with-chip value is now generic during configuration.
A default of generic for the choice of chip seems to make more sense
than none, this also matches with x86-64 which supports a 'generic'
chip, but not a 'none' chip.  The README file, also claims that a
straight ./configure on an x86-64 machine, should result in a successful
build, which it does with this patch.

ChangeLog:

	* configure.ac: When no --with-chip is specified, the default is
	now generic.
	* configure: Regenerate.