Regenerate BEEBS info file.
The info file doc/beebs.info is generated based on doc/beebs.texi, which
has recently changed, but the updated info file was not committed at the
same time.

In this patch I regenerate the info file, and also the version.texi
file, the version that is currently committed is very old.

ChangeLog:

	* doc/beebs.info: Regenerate after recent commits that changed
	doc/beebs.texi.
	* doc/version.texi: Regenerate a more upto date version file.