Remove duplicate 'Suite' in BEEBS header comments.
When I fixed the expansion of BEEBS a while back, in some cases (lots of
the Makefile.am files) the word Suite was actually on the following
line.
When I automatically updated the expansion I failed to spot this, and
added the word Suite in a second time.

This patch I remove the duplicate word 'Suite' and regenerate the
Makefile.in files.

ChangeLog:

	* Makefile.am: Remove duplicate word 'Suite' in header.
	* Makefile.in: Regenerate.
	* analysis/Makefile.am: Remove duplicate word 'Suite' in header.
	* src/*/Makefile.am: Likewise.
	* src/*/Makefile.in: Regenerate.