Fix 'make distcheck'.
Get 'make distcheck' working again.  This is just a case of source
files missing from the SOURCES lists in various places, add the files
to the list and regenerate the makefiles.  The only exception is the
new chipsupport.c file, in this case the generic code places a
dependency on chipsupport.c without checking if the file exists.  For
simplicity I've just added an empty file.

ChangeLog:

	* config/x86_64/chips/generic/chipsupport.c: New file.
	* src/common.mk.am: Add support.h as a source file.
	* src/miniz/Makefile.am: Add miniz.h as a source file.
	* src/qrduino/Makefile.am: Remove exxtable.h, and add ecctable.h
	qrbits.h to the source file list.
	* src/sglib-rbtree/Makefile.am: Add sglib.h as a source file.
	* src/*/Makefile.in: Regenerate.