Initialise variables, and turn of strict-aliasing.
A few benchmarks cause gcc to raise strict aliasing warnings, which we
ask to have reported as errors (-Werror).  For these benchmarks I've
turned off strict-aliasing, rather than try to rewrite the code to avoid
the warnings.

There's a bunch of variable initialisation added to avoid more compiler
warnings.

In one case, ludcmp, I've increased the size of a global array.  The
compiler is warning that the code might access beyond the bounds of the
array.  Without spending more time looking at the code I can't tell if
the code is wrong, or just really smart and so we know that the accesses
are all fine.  Given that the sizes involved are all very small I've
just increased the sizes of the global arrays to prevent the warnings.
This is a HACK, but if we assume the code is correct then my change has
done nothing but waste a little memory.