Add example configuring for a board
It can be a little tricky to determine what's wrong if errors
appear when compiling for a board without a concrete example, so
these documentation changes add an example of the configure flags
used to build for the STM32F0DISCOVERY board.

Some extra directions are added to the README about configuring
for a board too - this saves having to successfully configure and
make to build the info page that explains how to compile for a
board.