Conditionally set expected crc32 result
It is understood that on x86 the expected result differs from
Cortex-M3. For now, we define the expected result based on whether
the system is x86 or not. In future, we should check whether there
is an issue in the benchmark - perhaps the result should be the
same across all architectures.