Resolve two reproducibility issues.
This commit fixes two issues that can impact reproducibility:

1. Maintainer mode is now disabled by default. Having maintainer
mode turned on can cause regeneration of autotools files at build
time, which could result in the build being different in subtle
and hard-to-understand ways between machines.

2. Move enablement of dummy libraries into the chip configuration.
Previously the settings used for configuration were split between
two separate sources - the chip configuration, and the flags
passed to configure. This burdens anyone trying to reproduce a
build by requiring them to pass the correct flags to configure for
each configuration. Having the enablement of dummy libraries in
the chip configurations means that there is now one canonical
source for settings used by a given configuration, which is the
chip.cfg file.

Additionally, this reduces the risk of not noticing when a local
build has been done different dummy library settings, because any
changes will now show up in "git diff" instead of being invisible,
as when configuring with a different / incorrect set of flags.

Finally, we add files for the nolibc-nolibgcc-nolibm
configuration, which otherwise didn't differ from the
nolibc-nolibgcc configuration before - the only difference between
the two previously was the flags passed to configure.

ChangeLog:

        * configure: Regenerate.
        * configure.ac: Disable maintainer mode by default. Replace
        --enable-dummy arguments with USE_DUMMY_ variables.
        * config/arc/chips/compare-nocrt/chip.cfg: Add USE_DUMMY settings.
        * config/arc/chips/compare-nolibc-nolibgcc-nolibm/chip.cfg: New file.
        * config/arc/chips/compare-nolibc-nolibgcc/chip.cfg: Add USE_DUMMY settings.
        * config/arc/chips/compare-nolibc/chip.cfg: Add USE_DUMMY settings.
        * config/arm/chips/compare-nocrt/chip.cfg: Add USE_DUMMY settings.
        * config/arm/chips/compare-nolibc-nolibgcc-nolibm/chip.cfg: New file.
        * config/arm/chips/compare-nolibc-nolibgcc/chip.cfg: Add USE_DUMMY settings.
        * config/arm/chips/compare-nolibc/chip.cfg: Add USE_DUMMY settings.
        * config/riscv32/chips/compare-nocrt/chip.cfg: Add USE_DUMMY settings.
        * config/riscv32/chips/compare-nolibc-nolibgcc-nolibm/chip.cfg: New file.
        * config/riscv32/chips/compare-nolibc-nolibgcc-nolibm/chipsupport.c: New file.
        * config/riscv32/chips/compare-nolibc-nolibgcc-nolibm/chipsupport.h: New file.
        * config/riscv32/chips/compare-nolibc-nolibgcc/chip.cfg: Add USE_DUMMY settings.
        * config/riscv32/chips/compare-nolibc/chip.cfg: Add USE_DUMMY settings.