Remove control flags for non-benchmark directories
Some of the directories under beebs/src/ don't contain benchmarks, but
instead contain code common to many benchmarks.

It doesn't make sense (I think) to have a control flag for the common
directory, so lets remove them.

OK, so we could have the control flag disable all of the dependant
benchmarks, but that would be more work to implement, so for now, lets
just remove the flag.

ChangeLog:

	* configure: Regenerate.
	* configure.ac: Remove unneeded control flags.
	* genmacros.sh: Only generate flags for directories that contain
	benchmark programs.