Made changes to the select and miniz benchmarks.
The select benchmark contains a macro that induces a warning. Due
to the -Werror parameter this warning breaks the build. All is
correct now. The tested arch is ARM.

The miniz benchmark has a very long line of C code that is
misleading and the -Werror parameter reports it as a warning.
Fixed now. The tested arch is ARM.