Fix build when ferror is defined as a macro
On some systems ferror is a macro. When this is the case,
implementing a function names ferror can result in a syntax error,
similar to:

```
dummy-libc.c:248:1: error: expected identifier or '(' before 'int'
 ferror (FILE *stream __attribute__ ((unused)))
 ^~~~~~
dummy-libc.c:248:1: error: expected ')' before '(' token
 ferror (FILE *stream __attribute__ ((unused)))
```

This commit attempts to remedy the issue by undefining ferror in
dummy-libc.c if it is defined.

ChangeLog:

        * support/dummy-libc.c: Undef ferror if it is defined.