Attempt to fix bugs #59, #60 and others.
To fix #59 (Configure script ignores CFLAGS etc. from board / chip
configuration) we add the board flags to flags before doing any config
checks.

To fix #60 (Including chipsupport.c in chip.c doesn't work), we include
config.h in chip.h, so that it has the appropriate setting of
HAVE_BOARDSUPPORT_H. This allows it to include chipsupport.c correctly
as necessary.

This then exposes a new issue - any board whose chipsupport.c contains
startup code (instead of using startup files provided by libc) cannot
link because main cannot be resolved due to the link order. Rather than
doing something complicated with groups when linking or attempting to
change link order to resolve this, we instead merge libmain and
libsupport into a single libsupport library.

The next issue that requires attention is that the configuration for the
Taiyo Yuden EYSGJN results in empty executables, as there appears to be
nothing explicitly linked in (no startup code is included for this
board). We resolve this by setting the vector table as the ELF entry
point, to force it and all the relocations in it to be included. The
board / loader doesn't use the ELF entry point, so this shouldn't result
in incorrect execution.

We also add a DejaGNU board file for the Taiyo Yuden EYSGJN board, so
that the size of executables built for the board can easily be assessed
with `make check`.