Add support for STM32F769I-DISCO board.
The ST Microelectronics' discovery board for STM32F769 uses an
ARM Cortex-M7 based target with lots of features. The board
itself has rich peripherals. Currently the toggle pin chosen
is PA4 (designated as Arduino A1 at the back of the board).
The operating frequency is 216MHz, I- and D-caches are on,
the FPU is on (during init and with the proper compiler
parameters) for maximum performance. The GPIO capability is
quite good - I can see ~43MHz toggle frequency on my oscillo-
scope if I do a toggle without delays.

An interesting thing to note is that the board cannot be programmed
with openocd v.0.10.0 but the AC6's SW4STM32 succeeds in doing so.
You just have to make an empty project with the correct chip
selected (STM32F769NIH) and select the BEEBs elf files from the
Debug configuration window.